MockSum modified

+++ src/main/controller/MockSum.java
@@ -1,5 +1,5 @@
 package main.controller;
 
 public interface MockSum {
-    public double auxiliarSum(double x, double y);
+    public float auxiliarSum(double x, double y);
 }
